refactor(preinstall): improve disk confirmation logic for Clear-Disk operation

+++ OSDCloud/private/steps/3-preinstall/step-preinstall-cleartargetdisk.ps1
@@ -11,8 +11,13 @@ function step-preinstall-cleartargetdisk {
     $Step = $global:OSDCloudCurrentStep
     #=================================================
     #region Main
+    if ($global:OSDCloudDeploy.Force -eq $true) {
+        Write-Host -ForegroundColor DarkGray "[$(Get-Date -format s)] Force was specified. Clear-Disk confirmation prompts are disabled"
+        $Confirm = $false
+    }
+
     # If Confirm is set to false, we need to check if there are multiple disks
-    if (($Confirm -eq $false) -and (($global:OSDCloudWorkflowInvoke.GetDiskFixed | Measure-Object).Count -ge 2)) {
+    if (($Confirm -eq $false) -and ($global:OSDCloudDeploy.Force -ne $true) -and (($global:OSDCloudWorkflowInvoke.GetDiskFixed | Measure-Object).Count -ge 2)) {
         Write-Warning "[$(Get-Date -format s)] OSDCloud has detected more than 1 Fixed Disk is installed. Clear-Disk with Confirm is required"
         $Confirm = $true
     }
